Minor refactor of ScenarioRunner
